backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic[email]==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
passlib[argon2]==1.7.4  # Включает поддержку argon2
argon2-cffi==23.1.0     # Для argon2
httpx==0.25.0
